# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"University Timetable Admin API","version":"0.1.0"},"paths":{"/api/auth/login":{"post":{"summary":"Login For Access Token","operationId":"login_for_access_token_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/faculty/preferences":{"post":{"summary":"Submit Preferences","operationId":"submit_preferences_api_faculty_preferences_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreferenceSubmission"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/syllabus":{"get":{"summary":"Get Syllabus","operationId":"get_syllabus_api_syllabus_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/preferences":{"get":{"summary":"Get All Preferences","operationId":"get_all_preferences_api_admin_preferences_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/preferences/{pref_id}/approve":{"put":{"summary":"Approve Preference","operationId":"approve_preference_api_admin_preferences__pref_id__approve_put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"pref_id","in":"path","required":true,"schema":{"type":"integer","title":"Pref Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/preferences/{pref_id}/reject":{"put":{"summary":"Reject Preference","operationId":"reject_preference_api_admin_preferences__pref_id__reject_put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"pref_id","in":"path","required":true,"schema":{"type":"integer","title":"Pref Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/faculty-list":{"get":{"summary":"Get Faculty List","operationId":"get_faculty_list_api_admin_faculty_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/generate-workload":{"post":{"summary":"Generate Workload","description":"Takes an array of configured faculty configurations, runs the mathematical \nworkload calculation equation, checks for workload overflows, and saves \nthe result to the database before returning the validation status.","operationId":"generate_workload_api_admin_generate_workload_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/FacultyWorkloadConfig"},"type":"array","title":"Configs"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/generate-test":{"post":{"summary":"Generate Test Timetable","operationId":"generate_test_timetable_api_admin_generate_test_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/generate-batch":{"post":{"summary":"Generate Batch","operationId":"generate_batch_api_admin_generate_batch_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/timetable":{"get":{"summary":"Get Timetable","operationId":"get_timetable_api_admin_timetable_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/timetable/{block_id}":{"put":{"summary":"Update Timetable Block","operationId":"update_timetable_block_api_admin_timetable__block_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"block_id","in":"path","required":true,"schema":{"type":"integer","title":"Block Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditBlockRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/timetable/metadata":{"get":{"summary":"Get Metadata","operationId":"get_metadata_api_timetable_metadata_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/faculty/timetable":{"get":{"summary":"Get Faculty Timetable","operationId":"get_faculty_timetable_api_faculty_timetable_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/upload-metadata":{"post":{"summary":"Upload Metadata","operationId":"upload_metadata_api_admin_upload_metadata_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_metadata_api_admin_upload_metadata_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_upload_metadata_api_admin_upload_metadata_post":{"properties":{"syllabus_file":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Syllabus File"},"faculty_file":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Faculty File"},"rooms_file":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Rooms File"},"total_hours_file":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"Total Hours File"},"file":{"anyOf":[{"type":"string","contentMediaType":"application/octet-stream"},{"type":"null"}],"title":"File"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"file_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"File Type"}},"type":"object","title":"Body_upload_metadata_api_admin_upload_metadata_post"},"EditBlockRequest":{"properties":{"day":{"type":"integer","title":"Day"},"period":{"type":"integer","title":"Period"}},"type":"object","required":["day","period"],"title":"EditBlockRequest"},"FacultyWorkloadConfig":{"properties":{"faculty_id":{"type":"string","title":"Faculty Id"},"department":{"type":"string","title":"Department"},"theory_hours":{"type":"integer","title":"Theory Hours"},"lab_hours":{"type":"integer","title":"Lab Hours"},"incharge_hours":{"type":"integer","title":"Incharge Hours"},"max_hours_limit":{"type":"integer","title":"Max Hours Limit"}},"type":"object","required":["faculty_id","department","theory_hours","lab_hours","incharge_hours","max_hours_limit"],"title":"FacultyWorkloadConfig"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"PreferenceSubmission":{"properties":{"subjects":{"items":{"type":"string"},"type":"array","title":"Subjects"}},"type":"object","required":["subjects"],"title":"PreferenceSubmission"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}